Allow configuration of delay between replay steps
`Reanimator.replay` now accepts an optional config object as its second
argument. The `delay` key in this object sets the delay between steps when
replaying. It is a number or the string 'realtime'. If 'realtime' is specified,
Reanimator will make a good faith effort to replay the log with the captured
delay between events. If it is a number, Reanimator will wait that many ms
between steps. The default is 0.