Use native removeEventListener when replaying DOMContentLoaded asynchronously
Before, we used the wrapped version, which did not find `fire`, and thus could
not remove it. When `fire` ran, it triggered an infinite recursion by triggering
the event it was listening for.

Unfortunately, I haven't yet figured out how to test this one. DOMContentLoaded
is always fired before my call to `replay`.